LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbindex_keyPath.any-expected.txt


FAIL IDBIndex's keyPath attribute returns the same object. assert_equals: Same object instance is returned each time keyPath is inspected expected ["a", "b"] but got ["a", "b"]
